refactor(SaveButton): move the button position to editor header
